refactored ETH and BTC into coingeck_multibot -- will put all price bots there and remove individual scripts next
